feat: db model
